rosa_knob
This code was made by Mikolas Zuza for this project:
https://blog.prusa3d.com/3d-print-an-oversized-media-control-volume-knob-arduino-basics_30184/
Full credit goes to him. 

This is my own code implementation with the functions that I need it to do. Current functionality of the knob is as follows.

Rotary encoder:
left/right --> left/right arrow keys, mostly used as a cursor in text files
click --> toggles the shift button, useful to select text with great precision and decent speed
double-click --> disables shift state if active and deselects all selected text